+++ mne_bids/tests/test_utils.py
@@ -19,7 +19,8 @@ with warnings.catch_warnings():
                             category=ImportWarning)
     import mne
 
-from mne_bids import BIDSPath
+from mne.datasets import testing
+from mne_bids import BIDSPath, write_raw_bids
 from mne_bids.utils import (_check_types, _age_on_date,
                             _infer_eeg_placement_scheme, _handle_datatype,
                             _get_ch_type_mapping)
@@ -36,6 +37,29 @@ bids_path = BIDSPath(
     subject=subject_id, session=session_id, run=run, acquisition=acq,
     task=task)
 
+@pytest.fixture(scope='function')
+def return_bids_test_dir(tmpdir_factory):
+    """Return path to a written test BIDS dir."""
+    bids_root = str(tmpdir_factory.mktemp('mnebids_utils_test_bids_ds'))
+    data_path = testing.data_path()
+    raw_fname = op.join(data_path, 'MEG', 'sample',
+                        'sample_audvis_trunc_raw.fif')
+
+    event_id = {'Auditory/Left': 1, 'Auditory/Right': 2, 'Visual/Left': 3,
+                'Visual/Right': 4, 'Smiley': 5, 'Button': 32}
+    events_fname = op.join(data_path, 'MEG', 'sample',
+                           'sample_audvis_trunc_raw-eve.fif')
+
+    raw = mne.io.read_raw_fif(raw_fname)
+    raw.info['line_freq'] = 60
+    bids_path.update(root=bids_root)
+    # Write multiple runs for test_purposes
+    for run_idx in [run, '02']:
+        name = bids_path.copy().update(run=run_idx)
+        write_raw_bids(raw, name, events_data=events_fname,
+                       event_id=event_id, overwrite=True)
+
+    return bids_root
 
 def test_get_ch_type_mapping():
     """Test getting a correct channel mapping."""
@@ -124,3 +148,51 @@ def test_infer_eeg_placement_scheme():
     raw.rename_channels({'P3': 'foo'})
     placement_scheme = _infer_eeg_placement_scheme(raw)
     assert placement_scheme == 'n/a'
+
+def test_delete_scans(return_bids_test_dir, _bids_validate):
+    """Test delete scans in a dir."""
+    # deleting without subject, or session results in an error
+    bad_basename = make_bids_basename(task=task, run=run)
+    expected_err_msg = 'Deleting a scan requires the bids_basename '\
+                       'to have at least subject defined'
+    with pytest.raises(RuntimeError, match=expected_err_msg):
+        delete_scan(bad_basename, return_bids_test_dir)
+
+    # deleting without unique identifier results in an error
+    expected_err_msg = 'Deleting scan requires a unique bids_basename ' \
+                       'to parse in the scans.tsv file. '
+    with pytest.raises(RuntimeError, match=expected_err_msg):
+        delete_scan(bids_basename.copy().update(run=None),
+                    return_bids_test_dir)
+
+    # deleting scan should conform to bids-validator
+    delete_scan(bids_basename, bids_root=return_bids_test_dir)
+    _bids_validate(return_bids_test_dir)
+
+    # trying  to delete again would result in an error
+    with pytest.raises(RuntimeError, match='no files were found...'):
+        delete_scan(bids_basename, bids_root=return_bids_test_dir)
+
+    ses_path = op.join(return_bids_test_dir,
+                       f'sub-{subject_id}',
+                       f'ses-{session_id}')
+    scans_fpath = make_bids_basename(
+        subject=subject_id, session=session_id,
+        suffix='scans.tsv', prefix=ses_path)
+
+    # the scan for bids_basename should be gone inside
+    # scans.tsv and inside the actual session path
+    scans_tsv = _from_tsv(scans_fpath)
+    fnames = scans_tsv['filename']
+    assert all([str(bids_basename) not in fname for fname in fnames])
+
+    found_scans_fpaths = list(Path(ses_path).rglob(f'*{bids_basename}*'))
+    assert found_scans_fpaths == []
+
+
+def test_delete_participant(return_bids_test_dir, _bids_validate):
+    """Test delete participants."""
+    # deleting participant should conform to bids-validator
+    delete_participant(subject=subject_id, bids_root=return_bids_test_dir)
+    _bids_validate(return_bids_test_dir)
+    assert not op.exists(op.join(return_bids_test_dir, f'sub-{subject_id}'))